Move config to utils
