+++ Scripts/helmet_app.py
@@ -1,4 +1,4 @@
-from config import Config
+from utils.config import Config
 
 class HelmetApplication():
